+++ src/Lektra.hpp
@@ -18,7 +18,7 @@
 #include "TabWidget.hpp"
 #include "argparse.hpp"
 #ifdef ENABLE_LLM_SUPPORT
-#include "LLM/LLMWidget.hpp"
+    #include "LLM/LLMWidget.hpp"
 #endif
 #include "CommandManager.hpp"
 
@@ -214,6 +214,7 @@ private:
                           std::function<void()> onAllDone) noexcept;
     void focusSplitHelper(DocumentContainer::Direction direction) noexcept;
     void splitHelper(Qt::Orientation orientation) noexcept;
+    bool checkConfigFile(const QString &path) const noexcept;
 
     DocumentView *
     openFileSplitHelper(const QString &filename               = {},